bounding models with props
